requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
cobble==0.1.4
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
gunicorn==23.0.0
lxml==5.3.2
mammoth==1.9.0
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-docx==1.1.2
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
